+++ packages/gummiworm/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 2.0.11
+
+### Patch Changes
+
+- Updated dependencies [4cd67ec]
+  - @sundaeswap/core@2.10.0
+
+## 2.0.10
+
+### Patch Changes
+
+- Updated dependencies [2b8c9bf]
+  - @sundaeswap/core@2.9.0
+
 ## 2.0.9
 
 ### Patch Changes